KAFKA-17857 Move AbstractResetIntegrationTest and subclasses to tools
